implement getAllUserResults Api

+++ API/Dtos/ResultDto/SubmitResultDto.cs
@@ -4,7 +4,7 @@ public class SubmitResultDto
 {
     public string UserId { get; set; }
     public string ExamId { get; set; }
-    public int TimeTaken { get; set; }
+    public float TimeTaken { get; set; }
     public string Type { get; set; }
     public Dictionary<int, string> Answers { get; set; }
 }
