+++ API/Models/UserResult.cs
@@ -7,7 +7,7 @@ public class UserResult
     public string? ExamId { get; set; }
     public int Score { get; set; }
     public int CorrectAnswerAmount { get; set; }
-    public int TimeTaken { get; set; }
+    public float TimeTaken { get; set; }
     public DateTime CreatedAt { get; set; } = DateTime.UtcNow;
     public string? Type { get; set; }
     public List<DetailResult> DetailResults { get; set; }
